music_streaming_v2/requirements.txt
fastapi
strawberry-graphql[fastapi]
sqlalchemy[asyncio]
asyncpg
alembic
python-dotenv
uvicorn
python-multipart
python-dateutil
